requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
opencv-python-headless
albumentations
Pillow
python-multipart
